Service1/requirements.txt
alembic==0.9.9
aniso8601==3.0.0
Babel==2.5.3
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
docopt==0.6.2
Flask==0.12.2
Flask-API==1.0
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-DotEnv==0.1.1
Flask-HTTPAuth==3.2.3
Flask-JWT==0.3.2
Flask-JWT-Extended==3.8.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysql-connector==2.1.6
nose==1.3.7
passlib==1.7.1
pipreqs==0.4.9
pycparser==2.18
PyJWT==1.4.2
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
requests==2.18.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.6
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
yarg==0.1.9
